chore: rename to iterative upload
