+++ sn_cli/src/subcommands/files.rs
@@ -10,8 +10,8 @@ mod chunk_manager;
 mod estimate;
 
 pub(crate) mod download;
+pub(crate) mod iterative_upload;
 pub(crate) mod upload;
-pub(crate) mod upload_with_iter;
 
 pub(crate) use chunk_manager::ChunkManager;
